checkboxes reorganized
